requirements.txt
html2text
lxml
readability-lxml
requests
